devicetree: Make sure root node can be looked up via "/"
dt_find_node_by_path() implements the general rule that paths may not
end in a slash. That is true normally, but the special case of a path
that consists only of a slash is used to denote the root node. Add code
to make sure we treat that one right.

BRANCH=None
BUG=None
TEST=manual

Change-Id: Id5531a06e5d1f0859a9edf941c49d0ee7facedea
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1542702
Reviewed-by: Hung-Te Lin <hungte@chromium.org>