fit: Add device tree compression
This patch adds support for compressing individual device trees in the
FIT image. In order to make this efficient, we'll have to pull the
compatible property out of the FDT and store it directly in the config
node of the FIT image, so that we don't have to scan (and therefore
decompress) every single FDT on boot. Device tree compression is only
supported for FIT images that have this external compatible property.
For older images with no compression, we still support fallback to
scanning the FDT for the property.

BRANCH=None
BUG=None
TEST=Booted Kevin both with a legacy image and an image using the new
features.

Change-Id: I039feba5d0d625e329a884350134161864e9e86d
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1553458
Reviewed-by: Hung-Te Lin <hungte@chromium.org>