netboot: Disable video console after network transfer is finished
We enable the video console during netboot to make it easier for
developers to debug and distinguish network issues (e.g. no link, TFTP
server unreachable, file not found). The problem with the video console
is that once it reaches the end of the screen it starts to scroll, and
since the framebuffer is backed by writethrough memory memcpying the
whole thing line by line is very slow. We don't normally fill the screen
with just the netboot output, but the FIT loading output afterwards
generates a lot of spam (depending on the amount of packaged device
trees).

This is a quick and easy hack to disable further video console output
after we're through the network parts that really interest us. The last
thing users will see on a successful boot is the kernel command line.
Existing output will stay on the screen until the kernel takes over the
display, but it will not add new messages added by the boot code.

BRANCH=None
BUG=None
TEST=Booted Kevin.

Change-Id: I42b23f2d5b4e7dbef3dd7203655805740c7ab90f
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1544987
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>