sarien: Check for PD firmware in CBFS before adding
Vboot will fail if the PD firmware doesn't exist in CBFS, which
breaks builds that are not done with the chromium ebuilds.

To make this work again check if the PD hash exists in CBFS before
adding it as an update image.

BUG=none
BRANCH=none
TEST=boot without pdrw.bin/pdrw.hash

Change-Id: Ia17a327ca4375b82e594ec1e561529e5f1c7dfd1
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1572621
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>