depthcharge: only trust uart, gpio, and ec-based keyboards
Previously any non-USB keyboard was considered trusted.  With this
change only uart-based, gpio-based, and ec-based keyboards are trusted.
Ec-based keyboards are only trusted when a trusted ec is in use.

Cq-Depend: chromium:1614765
BUG=b:129471321
BRANCH=None
TEST=Local build and flash on Sarien and Atlas.  Verified that
keyboard was trusted on Atlas and was untruted on Sarien for
confirming the transition to developer mode.

Change-Id: Ie66c962ac0e99254651c2e3e3c03c96d65be0b71
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1614607
Reviewed-by: Matt Delco <delco@chromium.org>
Tested-by: John Su <john_su@compal.corp-partner.google.com>
Commit-Queue: John Su <john_su@compal.corp-partner.google.com>