kohaku depthcharge: skip SDHCI controller init
Kohaku does not have an SDHCI controller. The Hatch board setup tries
to initialize an SDHCI controller that is not present, resulting in an
error message in the depthcharge output. Modify board setup to first
check if the controller is present on the PCI bus, and only proceed
to initialization if it finds the controller.

BUG=b:132351898
BRANCH=None
TEST=build for Hatch reference and Kohaku with extra debug printf's
showing whether initialization is being skipped or not. Verify that
on Hatch, it initializes the SDHCI controller, and on Kohaku, it
recognizes that the controller is not present and skips the
initialization. Remove the debug printf's prior to commit.

Change-Id: I88a89860c92b6ea1e538e928dd3e365d62e1825e
Signed-off-by: Paul Fagerburg <pfagerburg@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1610254
Commit-Ready: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>