Makefile: Do not process linker script twice
CL:1600923 made some changes to linker script handling. Unfortunately,
by adding the linker script to the %.elf prerequisites, it also
(inadvertently) added it to the list of object files passed to LD. Turns
out that passing an .ldscript file as if it was another object tells LD
to use this script to "augment" the default script (even if it's exactly
the same file), and since redefining the same section has mostly no
effect the end result looks almost the same, except that some hardcoded
symbols and offsets are screwed up.

BRANCH=None
BUG=None
TEST=My Kevin boots again.

Change-Id: I82ce282d9e2d6f5908fb7018d720c7be78fd7435
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1616992
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Tested-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1645822
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Commit-Queue: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>