Auto-include Kconfig headers
The CONFIG/IS_ENABLED macros will assume that any option which is not
defined is 0 (disabled). This is dangerous for typos (which this patch
can't help) and also dangerous because if you forget to include
<config.h> all depthcharge Kconfigs will be considered disabled. Let's
switch the depthcharge build system to include Kconfig headers
automatically (like coreboot has already been doing for a long time) to
at least prevent the second kind of mistakes.

Also added a fix to make sure $(src) and $(obj) are absolute paths even
if they're passed in on the command line. This is important when passing
them to the subinvocation of $(MAKE) for vboot.

BRANCH=None
BUG=None
TEST=Booted Kevin.

Change-Id: I4a9225f25be158a57fb132d5ddd235aebef49cc0
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1629733
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1645823
Commit-Queue: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Tested-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>