mkbp: Remove MKBP_NO_INTERRUPT Kconfig
We don't need a Kconfig to decide whether we have a cros-ec interrupt
line registered... we can just make a tiny change to the
cros_ec_interrupt_pending() API to return that distinction to the
caller. This will allow x86 devices to register an interrupt line as
well if desired.

BRANCH=None
BUG=b:139384979
TEST=Booted Kevin, typed some keys and made sure they show up.

Change-Id: Ic17f7fcffd8d529681c4ec549f2470d8a1225adc
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1769243
Reviewed-by: Furquan Shaikh <furquan@chromium.org>