tpm: i2c: cr50: Retry i2c writes when the H1 nacks
When the H1 has been idle for several seconds it will nack the first
transaction on the i2c bus and if a second i2c transaction is started to
soon after the first it will be nacked as well. When this condition
happens in the call to request_locality() a -1 is returned and the final
result will be that the DUT does not boot and the recovery screen is
shown to the user.

This will ensure that the H1 is awake by retrying the i2c write up to 3
times. This matches the logic in the kernal cr50_i2c.c driver.

BUG=b:140748793
TEST=With a scope on the H1 i2c lines ensure that the DUT boots even if
     the first i2c transaction is nacked
BRANCH=none

Change-Id: Idb1f82e413603e6f1202c167aece55c57af8be88
Signed-off-by: Mathew King <mathewk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1817240
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>