board/drallion: Disable PD SW Sync on Drallion with board id 0
Currently PD FW is not expected to be updated on drallion boards with
id 0, while there will not be any issue with it on other Drallion boards.
This patch is intended to disable any PD firmware updates on boards with
id 0, by skipping ti_tcpc PD pointer initialization in depthcharge.

BUG=b:142629708
TEST=build and flash the coreboot image with GBB setting 0xb9, which by default
triggers EC and PD SW sync, since the test board is drallion with id 0 there is
no PD software sync from the depthcharge logs.

Change-Id: I64d39d8e8365fecb3860a5af65a60d4b96f1e1ab
Signed-off-by: Thejaswani Putta <thejaswani.putta@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1918633
Reviewed-by: Mathew King <mathewk@chromium.org>
Tested-by: Mathew King <mathewk@chromium.org>