drivers/usb: Always initialize USB explicitly
The current depthcharge USB framework is designed to initialize USB
on-demand the first time it is needed, although in practice it is always
initialized at the same point during input_enable(). With the new quirks
required for the AOA driver, we need more control over when exactly it
gets initialized... so continuing to pretend it is on demand is
disingenous and just risks that assumptions may break in the future.

This patch changes dc_usb_initialize() and its callers to make sure it
is only called at exactly one point (where we previously had
input_enable()). The DRIVER_INPUT_USB is thereby obsolete since its only
purpose was to initialize USB on demand... the actual usb_poll() on
havekey()/getchar() is already done by libpayload itself.

BRANCH=hatch,kukui
BUG=b:144970196
TEST=Booted Kevin in normal and developer mode, ensured USB was (not)
initialized in each as expected.

Change-Id: Idcf121081f061b5f1821f47b080e09f533d2ef5d
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1934960
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>