drivers/usb: Poll all devices twice (with short delay) on USB init
This patch tries to fix an issue observed with the AOA storage driver
that's only supposed to probe devices that weren't yet plugged in during
boot, where in certain cases a device seems to not get detected during
the first usb_poll(). The ugly but effective solution is to just poll
them all twice before we call the initialization "done", with an
arbitrarily selected delay in between.

This will increase boot times in developer and recovery mode across the
board by 300ms, which should hopefully be acceptable since we consider
those boots uncommon.

BRANCH=hatch,kukui
BUG=b:144970196
TEST=Booted Kevin

Change-Id: I5e47698136a6c0e88e80af0d37e7d8e2142b8d81
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1934961
Reviewed-by: Hung-Te Lin <hungte@chromium.org>