arm64: Don't warn for skipping low memory on kernels that support it
Newer kernels (since 4.6) seem to support being placed anywhere in DRAM
(as long as it's text_offset away from a 2MB boundary), and they
conveniently announce this support with a header flag. If we detect that
flag, we don't have to warn when we're skipping over a low memory range
that's too small to fit the kernel.

Also move the kernel header scratch buffer onto the stack because there
is no reason for it to be global.

BRANCH=None
BUG=b:145842053
TEST=None

Change-Id: I52dc5cc6a0b79585ed00e1b9c16a9418f03a77db
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/1956657
Reviewed-by: Stephen Boyd <swboyd@chromium.org>