ps8751: Also check for "broken" product IDs
Parade just let us know that the PS8805 reports a different product ID
(0x8803) when the firmware is in a broken state. We were only handling
the case where it reports all zeroes before, which was only true for
PS8751. This patch checks for the PS8805 case as well. (PS8815
supposedly keeps the same product ID so we should be fine there
already.)

BRANCH=nocturne?
BUG=b:148564462
TEST=None

Change-Id: I7d16f732dc928ade2aeb6290e6632fecfa4c36f5
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2031587
Reviewed-by: Douglas Anderson <dianders@chromium.org>