screens: Add warning banner for developer-signed firmware in recovery
We often hear that developers who are not deeply familiar with vboot get
confused and waste a lot of time understanding key mismatch issues when
trying to recover a device. In most cases, this is trying to run an
MP-signed recovery image on developer-signed firmware or vice versa, so
just being able to identify that distinction should hopefully help avoid
a lot of wasted time. (This information is technically already available
from the TAB screen, but for some reason people always look at me funny
when I recognize the developer root key hash from memory.)

Making changes to the official UI is a long and complicated process, but
thankfully developer-signed images are by definition not run by
end-users, so we don't really need to care that much about localization
and making it look nice for those. Therefore, this patch aims to solve
the problem by plastering a quick-and-dirty red warning banner across
the top of the screen on developer-signed images.

BRANCH=none
BUG=none
TEST=Booted on a Kevin and saw the banner.

Cq-Depend: chromium:2038245
Change-Id: Id49597618a9283bceeec80237cbafda8632d3da6
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2038246
Reviewed-by: Joel Kitching <kitching@chromium.org>