trogdor: Add PowerOps and correctly initialize sysinfo GPIOs
This patch registers the correct PowerOps for Trogdor so that we can
reboot and shut down from depthcharge, and it replaces the stubbed out
flag GPIOs by their real versions (from sysinfo or EC). Also switch the
SD card detect to the coreboot table which makes it easier to keep it in
sync.

BRANCH=None
BUG=None
TEST=Booted Trogdor

Change-Id: I29b6b25b6875fe54e4aab48dfb9cd0c9f7e13a58
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2044706