arm64: Do kernel handoff from C code
The arm64 kernel handoff is pretty simple and we have cleanly-defined
helpers for all the architectural things we need to do in libpayload.
There's no need to write a separate assembly function for it. (In
particular, the SCTLR masking in the old version was pretty hard to
read.)

This patch has the effect of leaving instruction cache and stack
alignment checking enabled during kernel handoff (which should be
legal).

BRANCH=None
BUG=None
TEST=Booted Trogdor

Cq-Depend: chromium:2060640
Change-Id: I8f806d88a3dd865c27bc80c7a717ebbc22428c15
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2057724