power: Reboot/power-off should never fail
The current implementations of cold_reboot() and power_off() can fail
and return an error, for example when a cleanup handler failed or the
platform driver had an issue. This seems dangerous -- some call sites
check the return value but others don't, and it seems that most people
assume these functions cannot return. In order to prevent unexpected
behavior, this patch changes their API to never return and instead die()
on error. This also means that run_cleanup_funcs() doesn't need to
return errors anymore (all we'd want to do is log them and it's easier
to do that within the function).

Also, the ec_reboot_to_ro() VbootEcOps callback implies that it reboots
immediately on success, but the current CrosEc version doesn't do that
and instead relies on the fact that call sites always call power_off()
afterwards anyway. Put the power_off() in the CrosEc driver so that it
matches the API expectation.

BRANCH=None
BUG=None
TEST=Booted Trogdor, did a dev mode transition.

Change-Id: I76e00bc99c99cd5be226e431d4d9e2e35cd5de19
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2068966
Reviewed-by: Joel Kitching <kitching@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>