aoa_recovery: Send custom "info block" to AOA device before MSC probe
This patch introduces an "AOA info block" that will be sent along the
BULK_OUT channel to the Android phone before transitioning to standard
USB mass-storage communication. This allows us to communicate some
custom information about the Chrome OS device to the phone. The
immediate motivation is to communicate the HWID, since Android 10 has a
bug that prevents AOA apps from reading the serial number descriptor
(which was previously used for this). While we're introducing this, also
add a few other bits that the phone might be interested in.

Since this is a protocol change, update the version number to 1.1.

BRANCH=None
BUG=b:153758197
TEST=None

Cq-Depend: chromium:2153592
Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I63cdf69d30a9d4cebf19b648b5a846da88b53444
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2151679
Reviewed-by: Jes Klinke <jbk@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>