deltaur: set FlashProtectionMapping for SPI flash.
A public version of the datasheet used is linked in the comments, for
posterity. Feel free to suggest I remove it if you think that's a bad
idea.

The wp_status_value is the same as Drallion, the previous generation. (In
fact, the datasheet is shared between Drallion's and Deltaur's chips.)

BUG=b:150165222
TEST=emerge-deltaur depthcharge
BRANCH=none

Change-Id: I5cf72f7ab43a0129f570784600b79124a1700002
Signed-off-by: Kevin Chowski <chowski@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2166560
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>