deltaur: fix syntax error.
I didn't upload the latest patchset to gerrit for the CL, and it got
through CQ because deltaur-cq doesn't seem to block submit when it
fails.

I also realized that I accidentally had expandtab on, so I swapped some
spaces for tabs. (I thought there was a presubmit for that, but maybe
not in depthcharge?)

BUG=b:150165222
TEST=emerge-deltaur depthcharge (this time on the patchset I uploaded.)
BRANCH=none

Change-Id: Ie479b8efe90595867ac4503cffd47b4d9fad7667
Signed-off-by: Kevin Chowski <chowski@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2169843
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>