drivers/bus/usb: add support for USB mux
This adds support for the concept of an SoC level mux for USB
ports. Tiger Lake is an example of an SoC that has an integrated USB
mux in front of some integrated USB ports which must be configured to
track external USB-C cable orientation.

The new weak function soc_usb_mux_init() is provided so that an SoC
mux subsystem can integrate with the USB subsystem.

BRANCH=none
BUG=b:149883933
TEST=tested indirectly by booting into a recovery image on USB-C

Change-Id: I6c96da5892706337804ddb92a91505331e8f1e6d
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2175991
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>