storage: aoa_recovery: Zero-initialize AOA info block
The AOA info block code never initializes the flags field to 0 (and it
seems that running memcpy() on any part of the structure makes the whole
thing invisible to -Wuninitialized). To fix this issue and make sure we
don't introduce new ones later, let's zero-initialize the whole
structure.

BRANCH=None
BUG=None
TEST=None

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I14a0bda3989b168572438cae2c4abcb180141ee4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2197947
Reviewed-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Jes Klinke <jbk@chromium.org>