fit: Swap compat matching priorities for board-revX and board-skuY
This patch changes the order in which compat strings involving revision
and SKU numbers are matched when looking for a compatible device tree.
The most precise match (board-revX-skuY) is still the highest priority,
but after that we will now first check for revision only (board-revX)
and then for SKU only (board-skuY). The reason for this is that SKU
differentiation is often added later to the project, so device trees for
earlier revisions may not have SKU numbers defined. So if we have a rev0
board (with sku0 as the "default SKU", because the board only started
having different SKUs with rev1) we want it to match the board-rev0
device tree, not board-sku0 which was added as an alias to
board-rev1-sku0 to provide the best known default for potential later
revisions of that SKU.

BRANCH=none
BUG=b:157234701
TEST=Observed "Compat preference:" console output change as expected.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I31ffadf9a6362f55a30f0a9154492c8db3ae15b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2212466
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>