depthcharge: Hide dev mode timeout description
In menu UI, when any key is pressed in developer mode screen, the
timer will be disabled. As a result, the description about
automatically booting after timeout should be removed.

Similarly, when the "Return to secure mode" button is hidden, the
description about it should be hidden.

In addition, fix argument type of error_code for headless devices.

BRANCH=none
BUG=b:146399181, b:157871585
TEST=emerge-puff depthcharge
TEST=USE="menu_ui" emerge-gale depthcharge

Cq-Depend: chromium:2241492
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Change-Id: Iec1d4a4d6b25414c92ac722152b0352230c7e7e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2236550
Reviewed-by: Joel Kitching <kitching@chromium.org>