net: add ramdisk argument to netboot()
While arm/arm64 have support for the FIT format which can include a
ramdisk separate from the main kernel image, other architectures such
as x86 only have support for plain kernel images (bzImage) in
Depthcharge.  Loading a ramdisk over TFTP therefore needs to be done
separately.  For this purpose, add a ramdisk argument to netboot()
with the name of an optional separate file to be downloaded as a
ramdisk and pass it on via the boot_info structure.

BRANCH=none
BUG=b:154536870
TEST=boot on x86

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>
Change-Id: I251df7a56589c0240761dd5caab174f535bf9ce9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2288612
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>