vboot/ui: Add in Language Menu on Broken/Invalid Screens
For some reason the Language menu was not included in the broken
screen, but the logic was there.  As a result the item highlighted and
the item was was selected did not match up.  Language menu was also
missing in the Recovery Invalid Screen (when you plug in an invalid
USB stick), so added there as well.

BUG=b:158776074,b:146399181
BRANCH=None
TEST=crossystem recovery_request=1
     Make sure Advanced Option is highlighted when first enter screen.
     Also, make sure that when we select the Advanced options menu and
     the Languages menu, the correct following menu is shown.  Boot
     into recovery, plug in an invalid USB and make sure resulting
     screen has a language menu displayed.

Signed-off-by: Shelley Chen <shchen@google.com>
Change-Id: I1b0e04f9e95b48afe0ba9c3636575ef8c82f03e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2242094
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>
Commit-Queue: Shelley Chen <shchen@chromium.org>