ps8751: add ps8815 A0 and A1 firmware update support
this updates the firmware update rules for the ps8815 A1 rev. now that
we can filter firmware blobs based on Device ID (chip rev.), the
firmware blob matching rules are updated to apply the correct firmware
update for the A0 and A1 revs of the ps8815 chips.

BUG=b:158857815
BRANCH=none
TEST=updated ps8815 A0 firmware to 0x03 and 0x04 and A1 from 0x06 to
	0x07.

Cq-Depend: chromium:2254992, chrome-internal:3130666
Change-Id: I29afcb23deb29933e9ec9d3ddb67d31531507a9d
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2255279
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>