vboot/ui: Retry with smaller width in ui_draw_bitmap()
When the "image exceeding canvas boundary" error is returned, instead of
giving up on the remaining drawing, try to draw the same bitmap again
with smaller width. Ideally this shouldn't happen. However, when we add
a string that is too long to fit in the canvas to bmpblk and don't
notice the length issue, this will be a nice fallback mechanism to avoid
drawing failure of the entire screen.

BRANCH=none
BUG=b:146399181, b:159399377
TEST=emerge-puff depthcharge

Change-Id: I8d9cb486156b49bdc7319f125e22cc21c7780822
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2288940
Reviewed-by: Joel Kitching <kitching@chromium.org>
(cherry picked from commit 8e83844b8f458618a9d60784667cffa8590c9e2b)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2308628
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Tested-by: Sam McNally <sammc@chromium.org>