drivers/storage/mmc: Increase clock speed before getting EXT_CSD
EXT_CSD is a 512 byte structure, or 4096 bits. In mmc_setup_media the
clock is set to the smallest possible value. The actual value depends on
the base clock frequency, but let's pick 100 KHz.

100000 bits/s * 1 s/1000ms = 100 bits/ms
4096 bits / (100 bits / ms) = 40ms

So we spend 40ms receiving EXT_CSD. The eMMC spec says that the CSD
contains the maximum clock speed supported for the current mode. This
change set the clock speed to the value specified in the CSD. This is
typically 25 or 26 MHz. This reduces the time to get EXT_CSD to <1ms.

BUG=b:158766134
BRANCH=none
TEST=Compare trembyle timestamps before and after this patch. Saw a
decrease of 45ms in trembyle boot.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I28bd74caf1c8ec4f158661ac69e4b673518a7f36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2321283
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@google.com>
Tested-by: Aaron Durbin <adurbin@google.com>