drivers/soc/picasso: Add eMMC HS400 workarounds
This change was inspired by the linux kernel. I took the approach of
being more verbose so it's easier to read and understand.
https://source.chromium.org/chromiumos/chromiumos/codesearch/+/master:src/third_party/kernel/v5.4/drivers/mmc/host/sdhci-acpi.c;l=548;drc=54b3c970e49abe32319deeae4b2a5933ca58653e

The eMMC DLL is disabled by default to save power. It needs to be
enabled before using HS400/DDR. Additionally the tuned clock can only be
used at the tuned frequency.

The HS400 tuning flow:
* Switch to HS200
* Tune
* Switch to HS
* Disable the tuning clock since we changed frequencies
* Switch to HS400
* Re-enable the tuning clock
* Enable the DLL

BUG=b:149101474
BRANCH=none
TEST=Build trembyle

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I217301192a18b461bd118fe18ce1f9e3c0608121
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2321611
Reviewed-by: Aaron Durbin <adurbin@google.com>