drivers/storage/mmc: Use slot_type to optimize eMMC setup
SD probing involves sending CMD8 followed by ACMD41. When the media is
eMMC, these commands are ignored resulting in the controller timing out
the operation. This results in 20ms spent probing for an SD card when
using eMMC.

In order to eliminate the 20ms, we can use the slot_type to skip the
probing.

When the slot type is REMOVABLE I still probe for MMC in case of a
timeout. This is to handle the case where an MMC card is inserted into
a removable slot. THough I don't think that's currently possible on any
of our devices.

BUG=b:158766134
BRANCH=none
TEST=Boot trembyle with tracing enabled and verified SD commands were
not set when booting from eMMC. Also plugged in an SD card and saw the
card initialized.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: Ie4f8013b33c8c30cc262ef6402b79215212e314f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2330371
Reviewed-by: Kangheui Won <khwon@chromium.org>