vboot/ui: Fix width of phone recovery QR code
After the auto linebreak functionality is introduced in CL:2303209,
the number of lines for the description text in the
VB2_SCREEN_RECOVERY_PHONE_STEP2 screen might change in different
locales. Therefore, fix the size of the QR code on that screen and set a
maximum width for the description in bmpblk to ensure that the text
won't overlap with the QR code.

BRANCH=puff
BUG=b:146399181, b:164943142
TEST=emerge-puff depthcharge
TEST=Description text didn't overlap with QR code for all languages

Cq-Depend: chromium:2362424
Change-Id: I413c59cd34860279113ee8a01b035000589f3b5f
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2364394
(cherry picked from commit 11f32cf5735c2231886344721418d00bdf139694)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2371465
Reviewed-by: Joel Kitching <kitching@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2376169
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>