base/list: Fix potential NULL pointer optimization problem
This patch is a direct port from a fix to the coreboot version of this
code (CB:44573). Apparently some newer versions of GCC more aggressively
assume that the address of a struct member can never be NULL. This list
iteration code relies on exactly that happening due to container_of()
shenanigans. Rewriting the comparison to use integers will prevent the
optimization.

BRANCH=None
BUG=None
TEST=None

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: If8c6538eb1c1eaaede5d0793541009e6f9060dab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2365772
Reviewed-by: Hung-Te Lin <hungte@chromium.org>