vboot: support buffered UI
Enable the buffered graphics feature from libpayload. This removes
flickering but adds input lag because the same amount of redrawing
(plus an extra memcpy, ~1.2ms on x86/puff) still has to be done and
is merely hidden from the user. It is not an excuse to stop
optimizing out unnecessary redraws, but in cases where that is not
possible it may provide a slightly better user experience.

BRANCH=zork
BUG=None
TEST=emerge-puff libpayload depthcharge chromeos-bootimage
     Builds and boots new firmware, seeing non-flickering UI
     (check the scroll bar in language selection UI).

Cq-Depend: chromium:2383657
Change-Id: I4106129a36e34590e0022ebb4e6c8af49b0134e0
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2386896
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>