drivers/storage/mmc: Add card_driver_strength callback
This callback allows the controller driver to specify which card drive
strength should be used by the card. The SDHCI driver will be modified
to query the SDHCI preset registers to determine the drive strength.

I also converted the MMC_TIMING_X defines into an enum. This makes it
clear what data type we are passing around.

This change does not support DDR52. mmc_select_ddr52 is using
mmc_select_hs to set the card timing. This needs to be refactored.

This change is a NOP. The system was implicitly using driver strength B.

BUG=b:159823235, b:162808954
BRANCH=zork
TEST=Verified ezkinil still boots at HS400

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I1cf9572da2f137f7d49cdb0c323129c2049306f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2402344
Reviewed-by: Furquan Shaikh <furquan@chromium.org>