drivers/storage/sdhci: Implement card_driver_strength
This change introduces SDHCI_PLATFORM_VALID_PRESETS. It will signal to
the SDHCI driver that the presets are valid for the platform.

This CL allows using the presets to correctly set the cards driver
strength. If SDHCI_PLATFORM_VALID_PRESETS we default to driver strength
B which is the same default used in mmc.c.

BUG=b:159823235, b:162808954
BRANCH=zork
TEST=Verified card drive strength is now set to A on Ezkinil. Tried
changing the preset drive strength to B and verified card had drive
strength of B.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: Idd6a39c31a7328fc1b1f7faf42950a945d525959
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2404332
Reviewed-by: Furquan Shaikh <furquan@chromium.org>