drivers/soc/picasso: Set SDHCI_PLATFORM_VALID_PRESETS
The presets set by AGESA are not completely valid yet, so we can't
enable the Preset Value Enable bit on the SDHCI controller.
depthcharge doesn't currently do this, so it's OK to set the flag. This
will cause the card to use Driver Strength A for both HS200 and HS400.
This allows Vilboz to boot using HS400.

BUG=b:159823235, b:162808954
BRANCH=zork
TEST=Verified Ezkinil can boot with HS400, HS200, DDR52 and HS. Also
verified the driver strengths in each mode.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I45fd40080268f6e3dad0daee371e98cd6730a54f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2404335
Reviewed-by: Aaron Durbin <adurbin@google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>