drivers/storage/sdhci: Set controller driver strength using presets
If SDHCI_PLATFORM_VALID_PRESETS is set, the controller driver strength
is set using the value specified in the preset registers.

This change does introduce a behavior change. The driver type was never
getting cleared in the ctrl_2 register. So once it was set it was never
cleared. This means that during a transition like HS400 setup
(HS200->HS->DDR52->HS400), the driver strength got set to A and that was
used for both HS and DDR52 modes. This will only impact Zork since it's
the only board using HS400.

Why not just enable presets via ctrl_2? There is still some refactoring
that needs to happen in the mmc driver to skip setting clocks.

BUG=b:159823235, b:168211204
BRANCH=zork
TEST=Boot ezkinil with verify card and controller now use the same
driver strength.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: Ife1b78127ca9fb9963480a020eee2df80e8f998a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2404334
Reviewed-by: Aaron Durbin <adurbin@google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>