drivers/storage/sdhci: Fix confusing VDD_180 logic
The VDD_180 flag was being toggled depending on the timing. This doesn't
actually make sense for eMMC. eMMC has a fixed signaling voltage that is
determined by the platform, so toggling it depending on the timing is
wrong. SD has a default signaling voltage of 3.3V, and a well defined
protocol for switching to 1.8V signaling. Switching to 1.8V is only
required when using UHS-I. Since depthcharge doesn't currently support
UHS-I SD timings I moved the 1.8V signaling logic to sdhci_init.

I added SDHCI_PLATFORM_EMMC_33V_VCCQ in case there is a platform that
uses a 3.3V signaling voltage. I don't thing we have any of these
platforms since they would be limited to HS. I added the platform_info so
it's easy to fix any platforms that are broken by this change.

I also renamed SDHCI_CTRL_VDD_180 to SDHCI_CTRL_180V_SIGNALING_ENABLE so
that it matches the SDHCI spec. The flag has nothing to do with VDD.

BUG=b:159823235
BRANCH=zork
TEST=Verified on Ezkinil still boots in HS400 and HS.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I10acf06c6efdc65e083f68ba104f722d9479d704
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2406187
Reviewed-by: Aaron Durbin <adurbin@google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>