drivers/storage/mmc: Make HS400ES use correct DDR driver strength
DDR52 uses the HS timing on the card, but uses the bus width to specify
DDR. This means we can't use mmc_select_hs to set the proper card driver
strength. This mode is only used briefly while transitioning, but I'm
going to make mmc_select_hs select the bus width so it matches the other
mmc_select_x functions, so I want to remove the other callers.

BUG=b:159823235
BRANCH=zork
TEST=none

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I76a8015c77c9b004d7b5fc47d5992d510ea74916
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/2412879
Reviewed-by: Aaron Durbin <adurbin@google.com>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>